+++ src/triangle.h
@@ -9,10 +9,10 @@ struct Triangle {
   std::array<Vector, 3> points;
 
   struct Intersection {
-    double t = std::numeric_limits<double>::infinity();
-    double epsilon = 0.0;
-    double b1 = 0.0;
-    double b2 = 0.0;
+    float t = std::numeric_limits<float>::infinity();
+    float epsilon = 0.0;
+    float b1 = 0.0;
+    float b2 = 0.0;
   };
 };
